healthomics_tools/v2.0/requirements.txt
boto3==1.35.20
GitPython
